+++ packages/connectivity/src/index.ts
@@ -6,7 +6,6 @@
 
 export {
   parseDestination,
-  noDestinationErrorMessage,
   toDestinationNameUrl,
   sanitizeDestination,
   CachingOptions,