+++ packages/connectivity/src/scp-cf/resilience-options.ts
@@ -1,3 +1,6 @@
+/**
+ * Options to configure resilience when fetching destinations.
+ */
 export interface ResilienceOptions {
   /**
    * A boolean value that indicates whether to execute request to SCP-CF services using circuit breaker.