+++ packages/odata-common/src/uri-conversion/get-resource-path.ts
@@ -32,6 +32,7 @@ export function createGetResourcePathForKeys(
   uriConverter: UriConverter
 ): GetResourcePathForKeys {
   /**
+   * @internal
    * Get the resource path of an entity specified by key-value pairs.
    * @typeparam EntityT - Type of the entity to get the resource path for.
    * @param keys - Key-value pairs where the key is the name of a key property of the given entity and the value is the respective value.