+++ packages/odata-v2/src/de-serializers/payload-value-converter.ts
@@ -1,6 +1,13 @@
 import { EdmTypeShared } from '@sap-cloud-sdk/odata-common/internal';
 import { DeSerializers } from './de-serializers';
 
+/**
+ * Convert a value from EDM format to TypeScript. This function is typically used when deserializing an OData response.
+ * @param value - The original value with EDM format.
+ * @param edmType - The EDM type of that value.
+ * @param deSerializers - (De-)serializers used for transformation.
+ * @returns Deserialized value.
+ */
 export function edmToTs(
   value: any,
   edmType: EdmTypeShared<'v2'>,