+++ scripts/get-changelog.ts
@@ -1,7 +1,13 @@
-import { unixEOL } from '@sap-cloud-sdk/util';
-import { openFile, version } from './util';
+import { readFileSync } from 'fs';
+import { currentSdkVersion } from './current-sdk-version';
 
-export function getChangelog(v = version): string {
+const unixEOL = '\n';
+
+function openFile(filePath: string): string {
+  return readFileSync(filePath, { encoding: 'utf8' });
+}
+
+export function getChangelog(v = currentSdkVersion): string {
   const changelog = openFile('CHANGELOG.md');
   const [, olderLogs] = changelog.split(`${unixEOL}# ${v}`);
   const logs = olderLogs.split(`${unixEOL}# `)[0];