+++ packages/generator-common/CHANGELOG.md
@@ -1,3 +1,8 @@
-# Changelog
+# @sap-cloud-sdk/generator-common
 
-All notable changes to this module will be documented centrally in the [SAP Cloud SDK changelog](https://github.com/sap/cloud-sdk-js/CHANGELOG.md).
+## 2.4.0
+
+### Patch Changes
+
+- Updated dependencies [0a008674]
+  - @sap-cloud-sdk/util@2.4.0